restructure architecture, implement Tag component articleCard, mainArticle
