+++ foolproject/src/components/articlePage/Ticker.vue
